refactor: layout project
