+++ songbird/src/Components/AudioPlayer/audioPlayer.js
@@ -1,5 +1,8 @@
+import './AudioPlayer.scss';
+
 const playAudio = (audioLink) => {
     const audio = new Audio(audioLink);
     audio.play()
 };
+
 export default  playAudio;
