+++ songbird/src/constants.js
@@ -1,8 +1,11 @@
 
-const audioCorrect = 'https://raw.githubusercontent.com/rolling-scopes-school/tasks/master/tasks/rslang/english-for.kids.data/audio/correct.mp3'
-const audioError = 'https://raw.githubusercontent.com/rolling-scopes-school/tasks/master/tasks/rslang/english-for.kids.data/audio/error.mp3'
+const AUDIO_CORRECT = 'https://raw.githubusercontent.com/rolling-scopes-school/tasks/master/tasks/rslang/english-for.kids.data/audio/correct.mp3'
+const AUDIO_ERROR = 'https://raw.githubusercontent.com/rolling-scopes-school/tasks/master/tasks/rslang/english-for.kids.data/audio/error.mp3'
 
-const audioSuccess = 'https://raw.githubusercontent.com/rolling-scopes-school/tasks/master/tasks/rslang/english-for.kids.data/audio/success.mp3'
-const audioFailure = 'https://raw.githubusercontent.com/rolling-scopes-school/tasks/master/tasks/rslang/english-for.kids.data/audio/failure.mp3'
+const AUDIO_SUCCESS = 'https://raw.githubusercontent.com/rolling-scopes-school/tasks/master/tasks/rslang/english-for.kids.data/audio/success.mp3'
+const AUDIO_FAILURE = 'https://raw.githubusercontent.com/rolling-scopes-school/tasks/master/tasks/rslang/english-for.kids.data/audio/failure.mp3'
 
-export { audioCorrect, audioError, audioSuccess, audioFailure };
+const DEFAULT_IMAGE = "https://i2.wp.com/pro-kletochki.ru/wp-content/uploads/2018/01/%D0%A7%D0%B5%D1%80%D0%BD%D0%BE-%D0%B1%D0%B5%D0%BB%D0%B0%D1%8F-%D1%80%D0%B0%D1%81%D0%BF%D0%B5%D1%87%D0%B0%D1%82%D0%BA%D0%B0-%D0%B4%D0%BB%D1%8F-%D0%9B%D0%94-%D0%A1%D0%BE%D0%B2%D0%B0.gif?w=365&h=433&ssl=1"
+const DEFAULT_NAME = '*******';
+
+export { AUDIO_CORRECT, AUDIO_ERROR, AUDIO_SUCCESS, AUDIO_FAILURE, DEFAULT_IMAGE, DEFAULT_NAME };
